[chore] base classes for generics [fix] config conflicts
