+++ lib/defaults/EmailDefault.ts
@@ -1,3 +1,11 @@
-export class EmailDefault {
+import { Generic } from '../Generic'
 
+export class EmailDefault extends Generic {
+  send(data) {
+    return Promise.resolve([])
+  }
+
+  sendTemplate(data) {
+    return Promise.resolve([])
+  }
 }
