+++ lib/defaults/ImageDefault.ts
@@ -1,0 +1,5 @@
+import { Generic } from '../Generic'
+
+export class ImageDefault extends Generic {
+
+}
