+++ lib/defaults/RenderDefault.ts
@@ -1,0 +1,10 @@
+import { Generic } from '../Generic'
+
+export class RenderDefault extends Generic {
+  render(data) {
+    return Promise.resolve({document: data})
+  }
+  renderSync(data) {
+    return { document: data }
+  }
+}
