+++ lib/defaults/index.ts
@@ -1,0 +1,9 @@
+export { DataStoreDefault } from './DataStoreDefault'
+export { EmailDefault } from './EmailDefault'
+export { FulfillmentDefault } from './FulfillmentDefault'
+export { GeolocationDefault } from './GeolocationDefault'
+export { ImageDefault } from './ImageDefault'
+export { PaymentDefault } from './PaymentDefault'
+export { RenderDefault } from './RenderDefault'
+export { ShippingDefault } from './ShippingDefault'
+export { TaxDefault } from './TaxDefault'
