+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@fabrix/spool-generics",
-  "version": "1.1.1",
+  "version": "1.1.2",
   "description": "Spool - Generic API",
   "homepage": "https://fabrix.app",
   "author": {
@@ -40,7 +40,7 @@
     "joi": "^13.4.0"
   },
   "devDependencies": {
-    "@fabrix/fabrix": "^1.1.3",
+    "@fabrix/fabrix": "^1.1.4",
     "@fabrix/lint": "^1.0.0-alpha.3",
     "@fabrix/spool-router": "^1.1.3",
     "@types/lodash": "^4.14.109",
@@ -57,7 +57,7 @@
     "typescript": "~2.8.1"
   },
   "peerDependencies": {
-    "@fabrix/fabrix": "^1.1.3",
+    "@fabrix/fabrix": "^1.1.4",
     "@fabrix/spool-router": "^1.1.3"
   },
   "engines": {
