+++ test/fixtures/FakeEmail.js
@@ -1,12 +1,6 @@
 'use strict'
-// const _ = require('lodash')
-const Generic = require('../../dist/Generic').Generic
-module.exports = class FakeEmailProvider extends Generic {
-  send(data) {
-    return Promise.resolve([])
-  }
 
-  sendTemplate(data) {
-    return Promise.resolve([])
-  }
+const Generic = require('../../dist/defaults').EmailDefault
+module.exports = class FakeEmailProvider extends Generic {
+
 }
