+++ test/fixtures/FakeGeolocation.js
@@ -1,11 +1,6 @@
 'use strict'
 // const _ = require('lodash')
-const Generic = require('../../dist/Generic').Generic
+const Generic = require('../../dist/defaults').GeolocationDefault
 module.exports = class FakeGeolocationProvider extends Generic {
-  locate(data) {
-    return Promise.resolve({
-      latitude: 0.000000,
-      longitude: 0.000000
-    })
-  }
+
 }
