+++ test/fixtures/FakeImage.js
@@ -1,6 +1,6 @@
 'use strict'
 // const _ = require('lodash')
-const Generic = require('../../dist/Generic').Generic
+const Generic = require('../../dist/defaults').ImageDefault
 module.exports = class FakeImageProcessor extends Generic {
 
 }
