+++ test/fixtures/FakeRender.js
@@ -1,11 +1,6 @@
 'use strict'
 // const _ = require('lodash')
-const Generic = require('../../dist/Generic').Generic
+const Generic = require('../../dist/defaults').RenderDefault
 module.exports = class FakeRenderService extends Generic {
-  render(data) {
-    return Promise.resolve({document: data})
-  }
-  renderSync(data) {
-    return { document: data }
-  }
+
 }
