+++ test/fixtures/FakeShipping.js
@@ -1,14 +1,6 @@
 'use strict'
 // const _ = require('lodash')
-const Generic = require('../../dist/Generic').Generic
+const Generic = require('../../dist/defaults').ShippingDefault
 module.exports = class FakeShippingProvider extends Generic {
-  validateAddress(data) {
-    return Promise.resolve(data)
-  }
-  getRate(data) {
-    return Promise.resolve({})
-  }
-  getRates(data) {
-    return Promise.resolve([])
-  }
+
 }
