+++ test/unit/generics/config.test.js
@@ -4,6 +4,6 @@ const assert = require('assert')
 
 describe('FakeGeneric', () => {
   it('should exist', () => {
-    assert.equal(global.app.config.get('foo'), 'bar')
+    assert.equal(global.app.config.get('fake_generic.foo'), 'bar')
   })
 })
